_legacy_python/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
PyJWT
psycopg2-binary
